토큰 ComposeUniqueKeyGenerator 인젝션 수정, AuthorizationCode IndexOutOfBoundsException 해결, 순환 참조 버그 수정

+++ src/main/kotlin/cube8540/oauth/authentication/oauth/token/infra/TokenIdGenerators.kt
@@ -37,16 +37,14 @@ class DefaultAuthorizationCodeGenerator(private val keyLength: Int): Authorizati
     override fun generate(): String {
         val bytes = ByteArray(keyLength)
         random.nextBytes(bytes)
-
         return getRandomCode(bytes)
     }
 
     private fun getRandomCode(bytes: ByteArray): String {
         val chars = CharArray(bytes.size)
         for (i in bytes.indices) {
-            chars[i] = DEFAULT_CODEC[(bytes[i] and 0xFF.toByte()) % DEFAULT_CODEC.size]
+            chars[i] = DEFAULT_CODEC[(bytes[i].toInt() and 0xFF) % DEFAULT_CODEC.size]
         }
-
         return String(chars)
     }
 }
